Refactor shipping rate validation and remove unneeded RateForm.
