+++ js/src/components/shipping-rate-section/estimated-shipping-rates-card/add-rate-form-modal.js
@@ -2,13 +2,14 @@
  * External dependencies
  */
 import { Button } from '@wordpress/components';
+import { Form } from '@woocommerce/components';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import RateForm from './rate-form.js';
 import RateModal from './rate-modal.js';
+import validate from './validate.js';
 
 /**
  * @typedef { import(".~/data/actions").CountryCode } CountryCode
@@ -31,7 +32,11 @@ const AddRateFormModal = ( {
 	onRequestClose,
 } ) => {
 	return (
-		<RateForm initialValues={ initialValues } onSubmit={ onSubmit }>
+		<Form
+			initialValues={ initialValues }
+			validate={ validate }
+			onSubmit={ onSubmit }
+		>
 			{ ( formProps ) => {
 				const { isValidForm, handleSubmit } = formProps;
 
@@ -59,7 +64,7 @@ const AddRateFormModal = ( {
 					/>
 				);
 			} }
-		</RateForm>
+		</Form>
 	);
 };
